// frontend/src/api/generated/models/correction-history-item.ts
/* tslint:disable */
/* eslint-disable */
/**
 * WMS ShowCase API
 * 倉庫管理システム（WMS）ShowCase のREST API定義
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface CorrectionHistoryItem
 */
export interface CorrectionHistoryItem {
    /**
     * 訂正日時
     * @type {string}
     * @memberof CorrectionHistoryItem
     */
    'correctedAt': string;
    /**
     * 前数量
     * @type {number}
     * @memberof CorrectionHistoryItem
     */
    'quantityBefore': number;
    /**
     * 後数量
     * @type {number}
     * @memberof CorrectionHistoryItem
     */
    'quantityAfter': number;
    /**
     * 訂正理由
     * @type {string}
     * @memberof CorrectionHistoryItem
     */
    'reason': string;
    /**
     * 実施者名
     * @type {string}
     * @memberof CorrectionHistoryItem
     */
    'executedByName': string;
}